Move Support Schools controller into the Providers module
