+++ app/views/support/providers/schools/delete.html.erb
